Move Echo bot to the end of the list
It looks like the Echo bot is slightly different from the rest in
that it just echoes whatever is said to it. Even though it could
be used to acquaint kids with the UI who are unfamiliar with android
devices or chat interfaces, it doesn't seem like anyone should
spend a lot of time with it unless needed, unlike other bots that
we would like kids to interact with and learn for relatively longer
durations.